requirements.txt
fastapi==0.109.1
uvicorn==0.23.1
PyYAML==6.0.1
pandas==2.2.2
numpy==2.0.0
pytz==2024.1
tzdata==2024.1
python-multipart==0.0.9